refactor: improve assets management for templates
